Fix some stuff
