+++ cmd/emi-wasm/main.go
@@ -1,3 +1,5 @@
+//go:build js && wasm
+
 package main
 
 import (
