+++ examples/vsql/.vscode/settings.json
