+++ examples/vsql/Makefile
@@ -1,4 +1,4 @@
-EMI ?= ../emi
+EMI ?= ../../emi
 
 generate:
 	$(EMI) go --path ./createuser.emi.yml --output ./dto --pkg dto
