+++ examples/vsql/dto/geopoint.go
