+++ examples/vsql/sqlfiles/embed.go
