+++ examples/vsql/sqlfiles/insert_data_types.sql
